ai_assessment/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
multiprocess==0.70.16
anyio==4.2.0
platformdirs==4.2.0
rich==13.7.0
pathlib==1.0.1
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.11.17
click==8.1.3
dataclasses-json==0.6.3
dnspython==2.3.0
exceptiongroup==1.1.1
filelock==3.9.0
fsspec==2023.12.2
grpcio==1.60.0
httpcore==0.16.3
httpx==0.24.1
idna==3.4
jsonpointer==2.1
kubernetes==28.1.0
marshmallow==3.19.0
overrides==7.4.0
posthog==3.1.0
prompt-toolkit==3.0.38
protobuf==4.25.1
Pygments==2.15.1
pytz==2023.3.post1
rapidfuzz==3.5.2
tqdm==4.65.0
urllib3==1.26.18
wcwidth==0.2.6
wrapt==1.14.0
azure-cognitiveservices-speech==1.36.0
pymongo==4.6.3
python-multipart==0.0.9
anthropic==0.25.7
openai==1.11.1
langchain==0.1.10
langchain-core==0.1.50
langchain-community==0.0.36
langchain_anthropic==0.1.11
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.5
mq_http_sdk==1.0.3
sqlmodel==0.0.14
mysql-connector==2.2.9
SQLAlchemy==2.0.24
tiktoken==0.5.2
pandas
pymysql